fibonacci

+++ functions_nested_loops/102-fibonacci.c
@@ -6,7 +6,7 @@
 
 int main(void)
 {
-	int count, i, j, k;
+	unsigned long count, i, j, k;
 
 	i = 0;
 	j = 1;
